move Parameters to parameters.py and clean up __init__
