+++ param/parameterized.py
@@ -50,7 +50,8 @@ from ._utils import (
     _recursive_repr,
     _validate_error_prefix,
     accept_arguments,
-    iscoroutinefunction
+    iscoroutinefunction,
+    descendents,
 )
 
 try:
@@ -364,25 +365,6 @@ def classlist(class_):
     Same as `inspect.getmro(class_)[::-1]`
     """
     return inspect.getmro(class_)[::-1]
-
-
-def descendents(class_):
-    """
-    Return a list of the class hierarchy below (and including) the given class.
-
-    The list is ordered from least- to most-specific.  Can be useful for
-    printing the contents of an entire class hierarchy.
-    """
-    assert isinstance(class_,type)
-    q = [class_]
-    out = []
-    while len(q):
-        x = q.pop(0)
-        out.insert(0,x)
-        for b in x.__subclasses__():
-            if b not in q and b not in out:
-                q.append(b)
-    return out[::-1]
 
 
 def get_all_slots(class_):
